refactor(events): decompose events.go into 4 files (<500 LOC each)
Split the 664-LOC events.go into:
- events.go (59 LOC): core types, interfaces, sentinel errors, constants
- event_bus.go (96 LOC): struct definition, constructor, options
- event_bus_lifecycle.go (220 LOC): lifecycle, shutdown, flush, listen
- event_bus_pubsub.go (319 LOC): publish, subscribe, dispatch

Pure structural change — no semantic modifications. All tests pass
with -race, go vet clean, verify_architecture shows zero new edges.
